moved fy-currency from add-edit-advance to shared
