+++ src/app/fyle/add-edit-advance-request/add-edit-advance-request.module.ts
@@ -8,9 +8,6 @@ import { AddEditAdvanceRequestPageRoutingModule } from './add-edit-advance-reque
 
 import { AddEditAdvanceRequestPage } from './add-edit-advance-request.page';
 import { MatIconModule } from '@angular/material/icon';
-import { FyCurrencyComponent } from './fy-currency/fy-currency.component';
-import { FyCurrencyChooseCurrencyComponent } from './fy-currency/fy-currency-choose-currency/fy-currency-choose-currency.component';
-import { FyCurrencyExchangeRateComponent } from './fy-currency/fy-currency-exchange-rate/fy-currency-exchange-rate.component';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { SharedModule } from 'src/app/shared/shared.module';
@@ -38,9 +35,6 @@ import { DraftAdvanceSummaryComponent } from './draft-advance-summary/draft-adva
   ],
   declarations: [
     AddEditAdvanceRequestPage,
-    FyCurrencyComponent,
-    FyCurrencyChooseCurrencyComponent,
-    FyCurrencyExchangeRateComponent,
     PolicyViolationDialogComponent,
     CameraOptionsPopupComponent,
     DraftAdvanceSummaryComponent,
