+++ src/app/fyle/add-edit-advance-request/add-edit-advance-request.page.html
@@ -57,6 +57,7 @@
             formControlName="currencyObj"
             [homeCurrency]="extendedAdvanceRequest.currency"
             [txnDt]="extendedAdvanceRequest?.created_at"
+            [componentInUse]="'advance-request'"
           >
           </app-fy-currency>
           <div
