--- src/app/fyle/add-edit-advance-request/fy-currency/fy-currency-choose-currency/fy-currency-choose-currency.component.html
@@ -1,51 +1,0 @@
-<ion-header mode="md">
-  <mat-icon class="fy-modal-top-bar" svgIcon="fy-rectangle"></mat-icon>
-  <ion-toolbar mode="md" class="fy-modal-toolbar">
-    <ion-title>
-      <div class="selection-modal--title">Select Currency</div>
-    </ion-title>
-    <ion-buttons slot="start">
-      <ion-button (click)="onDoneClick()">
-        <mat-icon class="fy-icon-close" svgIcon="fy-close"></mat-icon>
-      </ion-button>
-    </ion-buttons>
-  </ion-toolbar>
-  <ion-toolbar class="selection-modal--form-field-container">
-    <mat-form-field class="selection-modal--form-field mat-form-field-no-padding" appearance="none">
-      <input
-        #searchBar
-        id="select-currency-modal--search-input"
-        class="selection-modal--search-input"
-        matInput
-        placeholder="Search"
-        [(ngModel)]="value"
-      />
-      <button
-        mat-button
-        *ngIf="value"
-        matSuffix
-        mat-icon-button
-        aria-label="Clear"
-        class="selection-modal--clear"
-        (click)="clearValue()"
-      >
-        <mat-icon>close</mat-icon>
-      </button>
-    </mat-form-field>
-  </ion-toolbar>
-</ion-header>
-
-<ion-content>
-  <div class="selection-modal--sub-title">All Currencies</div>
-  <ng-container *ngFor="let currency of filteredCurrencies$ | async as list">
-    <div
-      matRipple
-      class="selection-modal--list-element"
-      [attr.id]="'currency' + currency.shortCode"
-      (click)="onCurrencySelect(currency)"
-    >
-      <div>{{ currency.shortCode }} - {{ currency.longName }}</div>
-      <mat-icon class="selection-modal--check" *ngIf="currency.shortCode === currentSelection"> check </mat-icon>
-    </div>
-  </ng-container>
-</ion-content>
