--- src/app/fyle/add-edit-advance-request/fy-currency/fy-currency-choose-currency/fy-currency-choose-currency.component.scss
@@ -1,52 +1,0 @@
-@import '../../../../../assets/styles/fy_select';
-
-.selection-modal {
-  color: black;
-  &--list-element {
-    padding: 18px 0;
-    margin: 0 16px;
-    @extend %fy-select-list-element-default-border-color-font;
-    @extend %fy-select-list-element-layout;
-  }
-
-  &--title {
-    @extend %fy-select-title;
-  }
-
-  &--header {
-    @extend %fy-select-header;
-  }
-
-  &--search-input {
-    @extend %fy-select-search-input;
-  }
-
-  &--form-field {
-    @extend %fy-select-mat-form-field;
-    .mat-form-field-wrapper {
-      box-shadow: rgba(0, 0, 0, 0.1) 0 4px 8px 0 !important;
-    }
-  }
-
-  &--form-field-container {
-    text-align: center;
-  }
-
-  &--sub-title {
-    color: $gray-3;
-    font-size: 14px;
-    font-weight: 700;
-    padding: 5px 14px;
-    border-bottom: 1px solid $grey-light;
-  }
-
-  &--clear {
-    background-color: $pure-white;
-    color: $black-light;
-    padding: 8px 5px;
-  }
-
-  &--check {
-    @extend %fy-select-check;
-  }
-}
