--- src/app/fyle/add-edit-advance-request/fy-currency/fy-currency-exchange-rate/fy-currency-exchange-rate.component.scss
@@ -1,45 +1,0 @@
-.exchange-rate-modal {
-  margin: 24px;
-  padding: 12px;
-  border: 1px solid #e5e5e5;
-  border-radius: 4px;
-  &--label {
-    margin: 0 30px 0 0;
-    font-size: 16px;
-    min-width: 48px;
-    color: #ababab;
-    font-weight: 500;
-    line-height: 1;
-  }
-
-  &--input {
-    font-size: 16px;
-    font-weight: 500;
-    line-height: 1;
-    height: 24px;
-    border: 0;
-  }
-
-  &--form-control {
-    padding: 12px;
-    border-bottom: 1px solid #e5e5e5;
-    display: flex;
-    align-items: center;
-    &:last-child {
-      border-bottom: 0;
-    }
-  }
-
-  &--description {
-    margin: 0;
-    color: #ababab;
-    text-align: center;
-    line-height: 1.5;
-    font-size: 16px;
-    padding: 0 16px;
-  }
-
-  &--save {
-    margin: 12px;
-  }
-}
