--- src/app/fyle/add-edit-advance-request/fy-currency/fy-currency.component.html
@@ -1,18 +1,0 @@
-<div class="fy-currency">
-  <form [formGroup]="fg" class="fy-currency--form">
-    <div class="fy-currency--input-container fy-currency--currency-block">
-      <div class="fy-currency--label">
-        Currency
-        <span class="fy-currency--mandatory"> * </span>
-      </div>
-      <input class="fy-currency--input" (click)="openCurrencyModal()" formControlName="currency" (blur)="onBlur()" />
-    </div>
-    <div class="fy-currency--input-container fy-currency--amount-block" [ngClass]="{ 'fy-currency__invalid': !valid }">
-      <div class="fy-currency--label">
-        Amount
-        <span class="fy-currency--mandatory"> * </span>
-      </div>
-      <app-fy-number placeholder="00.00" formControlName="amount"></app-fy-number>
-    </div>
-  </form>
-</div>
