--- src/app/fyle/add-edit-advance-request/fy-currency/fy-currency.component.scss
@@ -1,58 +1,0 @@
-@import '../../../../theme/colors.scss';
-
-.fy-currency {
-  &--label {
-    width: 100%;
-    margin-right: 2px;
-    font-size: 12px;
-    line-height: 16px;
-    font-weight: 400;
-    color: $black-light;
-  }
-
-  &--input {
-    font-size: 14px;
-    line-height: 18px;
-    color: $blue-black;
-    border: none;
-    width: 100%;
-    margin: 6px 0;
-    padding-left: 0;
-  }
-
-  &--mandatory {
-    color: $brand-primary;
-    display: inline-block;
-    font-size: 14px;
-    font-weight: 400;
-  }
-
-  &--input-container {
-    border-bottom: 1px solid $grey-lighter;
-  }
-
-  &--form {
-    display: flex;
-    align-items: center;
-    height: 70px;
-  }
-
-  &--amount {
-    font-size: 20px;
-    line-height: 1.2;
-    font-weight: 500;
-    margin-top: 0;
-  }
-
-  &--currency-block {
-    margin-right: 8px;
-  }
-
-  &--amount-block {
-    margin-left: 8px;
-  }
-
-  &__invalid {
-    border-bottom: 1px solid $red;
-  }
-}
