--- src/app/fyle/add-edit-advance-request/fy-currency/fy-currency.component.spec.ts
@@ -1,26 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { IonicModule } from '@ionic/angular';
-
-import { FyCurrencyComponent } from './fy-currency.component';
-
-xdescribe('FyCurrencyComponent', () => {
-  let component: FyCurrencyComponent;
-  let fixture: ComponentFixture<FyCurrencyComponent>;
-
-  beforeEach(
-    waitForAsync(() => {
-      TestBed.configureTestingModule({
-        declarations: [FyCurrencyComponent],
-        imports: [IonicModule.forRoot()],
-      }).compileComponents();
-
-      fixture = TestBed.createComponent(FyCurrencyComponent);
-      component = fixture.componentInstance;
-      fixture.detectChanges();
-    })
-  );
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
