+++ src/app/fyle/add-edit-expense/add-edit-expense.page.html
@@ -207,6 +207,7 @@
                   [txnDt]="fg?.controls?.dateOfSpend?.value"
                   [expanded]="attachedReceiptsCount === 0"
                   [disabled]="isUnifyCcceExpensesSettingsEnabled && isCccExpense"
+                  [componentInUse]="'expense-filing'"
                 >
                 </app-fy-currency>
                 <div
